Recents Ram Bar
Add Ram Bar to recents.

Notes.
1) You'll notice the Rambar gets updated frequently in the recents panel
for a few reasons to keep the data as up to date as possible, to update
after swiping away tasks, and also becuase this is now where we pick up
Boolean changes for if the bar is shown and if the all all button is
show.

2) You'll notice the data here does not match the data in
Settings->Apps-> Running. I used the same Math and got values from
the same place as settings but for some reason the data is still different
because the running process here is SystemUI and it has a larger foot print
then Settings.